+++ examples/base_cc_lb_zpa/main.tf
@@ -146,6 +146,7 @@ module "cc_vm" {
   ccvm_image_offer               = var.ccvm_image_offer
   ccvm_image_sku                 = var.ccvm_image_sku
   ccvm_image_version             = var.ccvm_image_version
+  ccvm_source_image_id           = var.ccvm_source_image_id
   cc_instance_size               = var.cc_instance_size
   mgmt_nsg_id                    = module.cc_nsg.mgmt_nsg_id
   service_nsg_id                 = module.cc_nsg.service_nsg_id
@@ -166,13 +167,14 @@ module "cc_vm" {
 #    created and assigned to ALL Cloud Connectors
 ################################################################################
 module "cc_nsg" {
-  source         = "../../modules/terraform-zscc-nsg-azure"
-  nsg_count      = var.reuse_nsg == false ? var.cc_count : 1
-  name_prefix    = var.name_prefix
-  resource_tag   = random_string.suffix.result
-  resource_group = module.network.resource_group_name
-  location       = var.arm_location
-  global_tags    = local.global_tags
+  source                 = "../../modules/terraform-zscc-nsg-azure"
+  nsg_count              = var.reuse_nsg == false ? var.cc_count : 1
+  name_prefix            = var.name_prefix
+  resource_tag           = random_string.suffix.result
+  resource_group         = module.network.resource_group_name
+  location               = var.arm_location
+  global_tags            = local.global_tags
+  support_access_enabled = var.support_access_enabled
 }