+++ modules/terraform-zscc-workload-azure/main.tf
@@ -84,7 +84,7 @@ resource "azurerm_linux_virtual_machine" "workload_vm" {
   resource_group_name = var.resource_group
 
   network_interface_ids = [azurerm_network_interface.workload_nic[count.index].id]
-  size                  = var.instance_size
+  size                  = contains(local.unsupported_regions, lower(var.location)) ? "Standard_A3" : var.instance_size
   admin_username        = var.server_admin_username
   computer_name         = "${var.name_prefix}-workload-${count.index + 1}-${var.resource_tag}"
   admin_ssh_key {
@@ -94,7 +94,7 @@ resource "azurerm_linux_virtual_machine" "workload_vm" {
 
   os_disk {
     caching              = "ReadWrite"
-    storage_account_type = "Premium_LRS"
+    storage_account_type = contains(local.unsupported_regions, lower(var.location)) ? "Standard_LRS" : "Premium_LRS"
   }
 
   source_image_reference {